Added the scope information into the MNEMONIC dictionary during the two passes.
Modified the way to determine DSECT symbols.

Added the address offset calculation in generating the report.
Added the ESD scope ID look-up table.
Modified the External Symbol Dictionary to record the ER type symbol as well as the SD type and PC type.
Added the Cross-Reference Sub-table to record the ER symbol (V-const).

Fixed a bug in parsing the length of a constant.

Fixed a serious problem in tokenlizing each line with regard to re splitting.

Modified the structure of the way op-code classes print out.
Included the op-code display in the report.